+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased](https://github.com/schemathesis/schemathesis/compare/v4.3.11...HEAD) - TBD
 
+### :rocket: Added
+
+- Adaptive stateful testing. It learns from response feedback to prioritize valid parameter combinations.
+
 ### :wrench: Changed
 
 - Infer links based on common REST patterns.
@@ -71,6 +75,7 @@
 
 - Setting an empty list to `phases.coverage.unexpected-methods` will disable testing of unexpected HTTP methods.
 - Schema bundler to correctly detect unbreakable indirect recursive cycles instead of failing with `RecursionError`.
+- Add matching required fields from known resources during dependency analysis.
 
 ### :wrench: Changed