+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Infer links based on common REST patterns.
 - Remove orphaned resources during dependency analysis.
 - Allow randomly skipping non-body links to test invalid scenarios.
+- Use `query`, `header` & `cookie` in dependency inference.
 
 ## [4.3.11](https://github.com/schemathesis/schemathesis/compare/v4.3.10...v4.3.11) - 2025-10-24